Complete Commons submodule extraction
