+++ Sources/BrowserServicesKit/Autofill/AutofillUserScript+SecureVault.swift
@@ -19,6 +19,7 @@
 
 import WebKit
 import os
+import Common
 
 // swiftlint:disable file_length
