+++ Sources/BrowserServicesKit/Autofill/AutofillUserScript.swift
@@ -20,6 +20,7 @@
 import WebKit
 import os.log
 import UserScript
+import Common
 
 public class AutofillUserScript: NSObject, UserScript {
