+++ Sources/BrowserServicesKit/Domain/AppVersion.swift
@@ -18,6 +18,7 @@
 //
 
 import Foundation
+import Common
 
 public struct AppVersion {
