+++ Sources/BrowserServicesKit/SmarterEncryption/HTTPSUpgradeParser.swift
@@ -18,6 +18,7 @@
 //
 
 import Foundation
+import Common
 
 public final class HTTPSUpgradeParser {
