+++ Sources/Common/Extensions/HashExtension.swift
